+++ x402_facilitator/test/x402_verify.test.js
@@ -4,8 +4,6 @@
 import { describe, test, expect, beforeEach, afterEach } from "vitest";
 import { verifyPayment } from "../x402_verify.js";
 import { resetFacilitator } from "../facilitator_instance.js";
-import { getChainConfig } from "../chain_utils.js";
-import { privateKeyToAccount, generatePrivateKey } from "viem/accounts";
 import { ExactEvmScheme } from "@x402/evm/exact/client";
 
 describe("x402 Verify", () => {
@@ -273,181 +271,7 @@ describe("x402 Verify", () => {
     expect(result.invalidReason).toBe("invalid_exact_evm_signature");
   });
 
-  /**
-   * END-TO-END SIGNATURE VALIDATION TEST
-   *
-   * This test creates a REAL signature using viem and x402 client libraries,
-   * then validates it through the full facilitator flow.
-   *
-   * This is critical because:
-   * 1. It tests the actual cryptographic signature verification
-   * 2. It catches signer configuration issues (like the toFacilitatorEvmSigner bug)
-   * 3. It serves as a reference implementation for TypeScript clients
-   *
-   * Unlike other tests that use mock signatures, this creates an authentic
-   * EIP-3009 authorization signature that the facilitator must validate.
-   *
-   * NOTE: uses a FRESH random key, not the shared Hardhat account #0. Since
-   * @x402/evm 2.17+ validates signatures via ERC-1271 `isValidSignature` when the
-   * signer address has code, and the well-known Hardhat key has a stray EIP-7702
-   * delegation on public testnets, that shared key takes the smart-wallet path and
-   * fails. A clean EOA (no code) uses ecrecover, as real users do.
-   */
-  test("E2E: validates real signature created with x402 client", async () => {
-    // CLIENT SIDE: Create a real signature using the client library
-    // This simulates what happens in a Jupyter notebook or web client
-
-    // Fresh random key => clean EOA (no code / no EIP-7702 delegation), the ecrecover path
-    const payerPrivateKey = generatePrivateKey();
-    const payerAccount = privateKeyToAccount(payerPrivateKey);
-
-    // Create x402 EVM client for signing
-    const evmClient = new ExactEvmScheme({
-      address: payerAccount.address,
-      signTypedData: async (args) => {
-        return payerAccount.signTypedData(args);
-      },
-    });
-
-    // Payment parameters. Source the USDC asset + EIP-712 domain name from the
-    // facilitator's own chain_utils (same source production uses via the 402
-    // response). Since @x402 2.17, verify derives the authoritative domain rather
-    // than trusting the payload's `extra`, so hardcoding a wrong name/address here
-    // would produce a domain mismatch (invalid_exact_evm_signature).
-    const network = "eip155:11155420"; // Optimism Sepolia
-    const cfg = getChainConfig(network);
-    const paymentAmount = "100000"; // $0.10 in 6-decimal USDC
-    const recipientAddress = "0x209693Bc6afc0C5328bA36FaF03C514EF312287C";
-
-    const paymentRequirements = {
-      scheme: "exact",
-      network,
-      amount: paymentAmount,
-      asset: cfg.USDC_ADDRESS,
-      payTo: recipientAddress,
-      maxTimeoutSeconds: 300,
-      extra: {
-        name: cfg.USDC_NAME,
-        version: "2",
-      },
-    };
-
-    // Create payment payload with REAL signature using x402 v2 API
-    const partialPayload = await evmClient.createPaymentPayload(2, paymentRequirements);
-
-    // Combine with resource information to create full payment payload
-    const paymentPayload = {
-      x402Version: 2,
-      resource: {
-        url: "https://api.example.com/e2e-test",
-        description: "End-to-end signature validation test",
-        mimeType: "application/json",
-      },
-      accepted: paymentRequirements,
-      payload: partialPayload.payload,
-    };
-
-    // FACILITATOR SIDE: Verify the signature
-    const result = await verifyPayment(paymentPayload, paymentRequirements);
-
-    // The signature validation should succeed, then fail at the balance check
-    // (invalid_exact_evm_insufficient_balance — the 2.17+ rename of insufficient_funds)
-    // because the fresh test wallet holds no USDC on Sepolia.
-    //
-    // What matters is that we DON'T get "invalid_exact_evm_signature", which would
-    // indicate a signature-verification / signer configuration problem.
-
-    if (!result.isValid) {
-      // Signature was validated correctly, but wallet doesn't have sufficient balance
-      // This is expected in test environment without real testnet funds
-      expect(result.invalidReason).toBe("invalid_exact_evm_insufficient_balance");
-      console.log("✓ Signature validation successful (insufficient funds is expected)");
-    } else {
-      // If the wallet happens to have funds, that's also valid
-      expect(result.isValid).toBe(true);
-      expect(result.invalidReason).toBeUndefined();
-      console.log("✓ Full payment verification successful");
-    }
-
-    // Verify the payer address is correctly identified
-    expect(paymentPayload.payload.authorization.from.toLowerCase()).toBe(
-      payerAccount.address.toLowerCase(),
-    );
-  });
-
-  test("validates signature for Optimism Mainnet (chainId 10)", async () => {
-    // This test would have failed before the multi-chain fix
-    // because the facilitator was hardcoded to use Sepolia client for all chains.
-    // Fresh random key => clean EOA (ecrecover path); see the E2E test note re:
-    // the shared Hardhat key's EIP-7702 delegation under @x402 2.17+.
-    const payerAccount = privateKeyToAccount(generatePrivateKey());
-
-    // Create a client for Optimism Mainnet (chainId 10)
-    const evmClient = new ExactEvmScheme({
-      address: payerAccount.address,
-      signTypedData: async (args) => payerAccount.signTypedData(args),
-    });
-
-    // Source asset + domain name from chain_utils. Note Optimism *mainnet* USDC
-    // is "USD Coin" (not "USDC"); hardcoding the wrong name is exactly what broke
-    // this test under @x402 2.17+, which derives the authoritative domain.
-    const network = "eip155:10"; // Optimism Mainnet
-    const cfg = getChainConfig(network);
-    const paymentAmount = "1000000"; // $1.00 in 6-decimal USDC
-    const recipientAddress = "0x209693Bc6afc0C5328bA36FaF03C514EF312287C";
-
-    const paymentRequirements = {
-      scheme: "exact",
-      network,
-      amount: paymentAmount,
-      asset: cfg.USDC_ADDRESS,
-      payTo: recipientAddress,
-      maxTimeoutSeconds: 300,
-      extra: {
-        name: cfg.USDC_NAME,
-        version: "2",
-      },
-    };
-
-    // Create payment payload with REAL signature for Mainnet (chainId 10)
-    const partialPayload = await evmClient.createPaymentPayload(2, paymentRequirements);
-
-    const paymentPayload = {
-      x402Version: 2,
-      resource: {
-        url: "https://api.example.com/mainnet-test",
-        description: "Mainnet multi-chain signature validation test",
-        mimeType: "application/json",
-      },
-      accepted: paymentRequirements,
-      payload: partialPayload.payload,
-    };
-
-    // CRITICAL TEST: Without the multi-chain fix, this would fail because
-    // verifyTypedData would use the Sepolia client instead of Mainnet client
-    // The domain.chainId would be 10 (Mainnet) but the client would be configured for 11155420 (Sepolia)
-    const result = await verifyPayment(paymentPayload, paymentRequirements);
-
-    // We accept insufficient-balance or unauthorized_agent — we're testing signature
-    // validation. The important part is that we DON'T get "invalid_exact_evm_signature",
-    // which would indicate the signer is using the wrong chain client.
-    if (!result.isValid) {
-      // Valid errors: no USDC balance, or agent not whitelisted on Mainnet
-      expect(["invalid_exact_evm_insufficient_balance", "unauthorized_agent"]).toContain(
-        result.invalidReason,
-      );
-      console.log(
-        `✓ Mainnet signature validation successful (${result.invalidReason} is expected)`,
-      );
-    } else {
-      expect(result.isValid).toBe(true);
-      console.log("✓ Mainnet full payment verification successful");
-    }
-
-    // Verify it's actually using mainnet chainId
-    expect(paymentPayload.accepted.network).toBe("eip155:10");
-  });
-
-  // Note: Additional blockchain integration tests (actual settlement on testnet)
-  // would require testnet funds and transaction execution
+  // Real-signature verification (live RPC) lives in the integration suite:
+  // test/integration/x402_verify_signature.integration.test.js — run via
+  // `npm run test:integration`. It's kept out of the hermetic unit run.
 });
